+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 - Update document. (2024-2-15)
 - Support report page performance data. (2024-3-10)
 - Support report performance data. (2024-3-10)
+- Support upload log info directly. (2024-4-17)
+- Support upload log info directly. (2024-4-17)
+- Support upload log info directly. (2024-4-17)
+- Support upload log info directly. (2024-4-17)